requirements.txt
annotated-types==0.7.0
asn1crypto==1.5.1
attrs==24.2.0
bech32==1.2.0
black==24.8.0
blockfrost-python==0.6.0
cachetools==5.5.0
cardano-tools==2.1.0
cbor2==5.6.4
certifi==2024.7.4
certvalidator==0.11.1
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
cose==0.9.dev8
cryptography==43.0.0
docker==7.1.0
ecdsa==0.19.0
ECPy==1.2.5
frozendict==2.4.4
frozenlist==1.4.1
humanfriendly==10.0
idna==3.7
mnemonic==0.21
mypy-extensions==1.0.0
ogmios==1.3.0
orjson==3.10.15
oscrypto==1.3.0
packaging==24.1
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.2.2
pprintpp==0.4.0
ptyprocess==0.7.0
pycardano @ git+https://github.com/Python-Cardano/pycardano.git@f21348f7619d8f7a750736480c0613848ec7a4e2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyNaCl==1.5.0
python-dotenv==1.0.1
requests==2.32.3
setuptools==72.2.0
six==1.16.0
typeguard==4.3.0
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
websockets==13.1
